requirements.txt
bpy
numpy
matplotlib
pillow
